requirements.txt
# 文献转 RAG 知识库 - 依赖

# PDF 和 Word 文档解析
pymupdf>=1.23.0
python-docx>=0.8.11

# 向量嵌入和检索
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.24.0